SERVER/pyproject.toml
[project]
name = "corporate-assistent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "bcrypt>=5.0.0",
    "certifi>=2026.1.4",
    "coroutine>=0.0.2",
    "fastapi>=0.128.0",
    "itsdangerous>=2.2.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-openai>=1.1.7",
    "motor>=2.4.0",
    "passlib>=1.7.4",
    "pinecone>=8.0.0",
    "pymongo[bson,srv]==3.11",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "streamlit>=1.52.2",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.40.0",
]
